pyproject.toml
[project]
name = "dryad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fire>=0.7.0",
    "gguf>=0.10.0",
    "jinja2>=3.1.4",
    "litestar>=2.12.1",
    "llama-cpp-python>=0.3.1",
    "msgspec>=0.18.6",
    "numpy>=2.1.3",
    "poethepoet>=0.30.0",
    "sentencepiece>=0.2.0",
    "uvicorn>=0.32.0",
]

[tool.poe.tasks]
serve = "python src/main.py"
